Velo/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
alembic==1.13.1
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.24.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4
python-multipart==0.0.9 